no message
git-svn-id: https://svn.grrrr.org/ext/trunk@706 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flsupport.h
@@ -92,7 +92,7 @@ public:
 		//! Free an aligned memory block
 		static void FreeAligned(void *blk);
 		//! Test for alignment
-		static bool IsAligned(void *ptr,int bitalign = 128)	{ return (reinterpret_cast<unsigned long>(ptr)&(bitalign-1)) == 0; }
+		static bool IsAligned(void *ptr,int bitalign = 128)	{ return (reinterpret_cast<unsigned long long>(ptr)&(bitalign-1)) == 0; }
 		
 	//!	@}  FLEXT_S_MEMORY
